+++ src/customServices.ts
@@ -18,6 +18,10 @@ export abstract class CustomServices {
 
 export abstract class CustomCharacteristics {
    static readonly AirPressureUUID: string = 'E863F10F-079E-48FF-8F27-9C2605A29F52';
+   static readonly VoltageUUID: string = 'E863F10A-079E-48FF-8F27-9C2605A29F52';
+   static readonly CurrentUUID: string = 'E863F126-079E-48FF-8F27-9C2605A29F52';
+   static readonly PowerUUID: string = 'E863F10D-079E-48FF-8F27-9C2605A29F52';
+   static readonly EnergyConsumptionUUID: string = 'E863F10C-079E-48FF-8F27-9C2605A29F52';
 
    static get AirPressure(): Characteristic {
      const characteristic = new hap.Characteristic('Air Pressure', CustomCharacteristics.AirPressureUUID);
@@ -27,6 +31,60 @@ export abstract class CustomCharacteristics {
        minValue: 700,
        maxValue: 1100,
        minStep: 1,
+     });
+     characteristic.value = characteristic.getDefaultValue();
+
+     return characteristic;
+   }
+
+   static get Voltage(): Characteristic {
+     const characteristic = new hap.Characteristic('Voltage', CustomCharacteristics.VoltageUUID);
+     characteristic.setProps({
+       format: hap.Formats.FLOAT,
+       perms: [hap.Perms.PAIRED_READ, hap.Perms.NOTIFY],
+       minValue: 0,
+       maxValue: 300,
+       // unit: Volt
+     });
+     characteristic.value = characteristic.getDefaultValue();
+
+     return characteristic;
+   }
+
+   static get Current(): Characteristic {
+     const characteristic = new hap.Characteristic('Current', CustomCharacteristics.CurrentUUID);
+     characteristic.setProps({
+       format: hap.Formats.FLOAT,
+       perms: [hap.Perms.PAIRED_READ, hap.Perms.NOTIFY],
+       minValue: 0,
+       maxValue: 16,
+       // unit: Amp
+     });
+     characteristic.value = characteristic.getDefaultValue();
+
+     return characteristic;
+   }
+
+   static get Power(): Characteristic {
+     const characteristic = new hap.Characteristic('Power', CustomCharacteristics.PowerUUID);
+     characteristic.setProps({
+       format: hap.Formats.FLOAT,
+       perms: [hap.Perms.PAIRED_READ, hap.Perms.NOTIFY],
+       minValue: 0,
+       maxValue: 10000,
+       // unit: Watt
+     });
+     characteristic.value = characteristic.getDefaultValue();
+
+     return characteristic;
+   }
+
+   static get EnergyConsumption(): Characteristic {
+     const characteristic = new hap.Characteristic('Energy Consumption', CustomCharacteristics.EnergyConsumptionUUID);
+     characteristic.setProps({
+       format: hap.Formats.FLOAT,
+       perms: [hap.Perms.PAIRED_READ, hap.Perms.NOTIFY],
+       // unit: kWh
      });
      characteristic.value = characteristic.getDefaultValue();