+++ CHANGELOG.md
@@ -6,6 +6,9 @@ As soon as the project reaches a mature and stable state, the first major versio
 and after the project will apply [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][]
+### Added
+
+- Electrical measurement information added for switches/lights that support it (using the Elgato Eve Energy characteristics).
 
 ## [0.0.7][] - 2020-10-15
 ### Added